+++ packages/zowe-explorer/src/Profiles.ts
@@ -41,6 +41,7 @@ import { ZoweExplorerApiRegister } from "./ZoweExplorerApiRegister";
 import { ZoweExplorerExtender } from "./ZoweExplorerExtender";
 import * as globals from "./globals";
 import * as nls from "vscode-nls";
+import { SettingsConfig } from "./utils/SettingsConfig";
 
 // Set up localization
 nls.config({
@@ -1075,7 +1076,7 @@ export class Profiles extends ProfilesCache {
 
         // Delete from Data Set Sessions list
         const dsSetting: any = {
-            ...vscode.workspace.getConfiguration().get(this.dsSchema),
+            ...SettingsConfig.getDirectValue(this.dsSchema),
         };
         let sessDS: string[] = dsSetting.sessions;
         let faveDS: string[] = dsSetting.favorites;
@@ -1087,11 +1088,11 @@ export class Profiles extends ProfilesCache {
         });
         dsSetting.sessions = sessDS;
         dsSetting.favorites = faveDS;
-        await vscode.workspace.getConfiguration().update(this.dsSchema, dsSetting, vscode.ConfigurationTarget.Global);
+        await SettingsConfig.setDirectValue(this.dsSchema, dsSetting);
 
         // Delete from USS Sessions list
         const ussSetting: any = {
-            ...vscode.workspace.getConfiguration().get(this.ussSchema),
+            ...SettingsConfig.getDirectValue(this.ussSchema),
         };
         let sessUSS: string[] = ussSetting.sessions;
         let faveUSS: string[] = ussSetting.favorites;
@@ -1103,11 +1104,11 @@ export class Profiles extends ProfilesCache {
         });
         ussSetting.sessions = sessUSS;
         ussSetting.favorites = faveUSS;
-        await vscode.workspace.getConfiguration().update(this.ussSchema, ussSetting, vscode.ConfigurationTarget.Global);
+        await SettingsConfig.setDirectValue(this.ussSchema, ussSetting);
 
         // Delete from Jobs Sessions list
         const jobsSetting: any = {
-            ...vscode.workspace.getConfiguration().get(this.jobsSchema),
+            ...SettingsConfig.getDirectValue(this.jobsSchema),
         };
         let sessJobs: string[] = jobsSetting.sessions;
         let faveJobs: string[] = jobsSetting.favorites;
@@ -1119,9 +1120,7 @@ export class Profiles extends ProfilesCache {
         });
         jobsSetting.sessions = sessJobs;
         jobsSetting.favorites = faveJobs;
-        await vscode.workspace
-            .getConfiguration()
-            .update(this.jobsSchema, jobsSetting, vscode.ConfigurationTarget.Global);
+        await SettingsConfig.setDirectValue(this.jobsSchema, jobsSetting);
 
         // Remove from list of all profiles
         const index = this.allProfiles.findIndex((deleteItem) => {
@@ -1893,8 +1892,9 @@ export class Profiles extends ProfilesCache {
     // Temporary solution for handling unsecure profiles until CLI team's work is made
     // Remove secure properties and set autoStore to false when vscode setting is true
     private createNonSecureProfile(newConfig: zowe.imperative.IConfig): void {
-        const configuration: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration();
-        const isSecureCredsEnabled: boolean = configuration.get(globals.SETTINGS_SECURE_CREDENTIALS_ENABLED);
+        const isSecureCredsEnabled: boolean = SettingsConfig.getDirectValue(
+            globals.SETTINGS_SECURE_CREDENTIALS_ENABLED
+        );
         if (!isSecureCredsEnabled) {
             for (const profile of Object.entries(newConfig.profiles)) {
                 delete newConfig.profiles[profile[0]].secure;