+++ packages/zowe-explorer/src/job/ZosJobsProvider.ts
@@ -28,10 +28,10 @@ import { getAppName, sortTreeItems, labelRefresh, jobPrefixValidator } from "../
 import { ZoweTreeProvider } from "../abstract/ZoweTreeProvider";
 import { getIconByNode } from "../generators/icons";
 import * as contextually from "../shared/context";
+import { resetValidationSettings } from "../shared/actions";
+import { SettingsConfig } from "../utils/SettingsConfig";
 
 import * as nls from "vscode-nls";
-import { resetValidationSettings } from "../shared/actions";
-import { PersistentFilters } from "../PersistentFilters";
 // Set up localization
 nls.config({
     messageFormat: nls.MessageFormat.bundle,
@@ -223,7 +223,7 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
      * @param {string} [profileType] - optional; loads profiles of a certain type if passed
      */
     public async addSession(sessionName?: string, profileType?: string) {
-        const setting = PersistentFilters.getDirectValue(globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION) as boolean;
+        const setting: boolean = SettingsConfig.getDirectValue(globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION);
         // Loads profile associated with passed sessionName, default if none passed
         if (sessionName) {
             const theProfile: imperative.IProfileLoaded = Profiles.getInstance().loadNamedProfile(sessionName);
@@ -527,7 +527,7 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
                 node.job,
                 node.getProfile()
             );
-            favJob.contextValue = node.contextValue;
+            favJob.contextValue = globals.JOBS_SESSION_CONTEXT + globals.FAV_SUFFIX;
             favJob.command = { command: "zowe.jobs.search", title: "", arguments: [favJob] };
             this.saveSearch(favJob);
         } else {
@@ -802,8 +802,6 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
             node.getSession().ISession.password = faveNode.getSession().ISession.password;
             node.getSession().ISession.base64EncodedAuth = faveNode.getSession().ISession.base64EncodedAuth;
         }
-        const jobQueryObj = this.parseJobSearchQuery(searchCriteria);
-        this.applySearchLabelToNode(node, jobQueryObj);
         return searchCriteria;
     }
 
@@ -821,6 +819,8 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
                 searchCriteria = await this.applyRegularSessionSearchLabel(node);
             } else {
                 searchCriteria = await this.applySavedFavoritesSearchLabel(node);
+                const jobQueryObj = this.parseJobSearchQuery(searchCriteria);
+                this.applySearchLabelToNode(node, jobQueryObj);
             }
             if (!searchCriteria) {
                 return undefined;
@@ -839,13 +839,13 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
 
     public async onDidChangeConfiguration(e: vscode.ConfigurationChangeEvent) {
         if (e.affectsConfiguration(ZosJobsProvider.persistenceSchema)) {
-            const setting: any = { ...vscode.workspace.getConfiguration().get(ZosJobsProvider.persistenceSchema) };
+            const setting: any = {
+                ...SettingsConfig.getDirectValue(ZosJobsProvider.persistenceSchema),
+            };
             if (!setting.persistence) {
                 setting.favorites = [];
                 setting.history = [];
-                await vscode.workspace
-                    .getConfiguration()
-                    .update(ZosJobsProvider.persistenceSchema, setting, vscode.ConfigurationTarget.Global);
+                await SettingsConfig.setDirectValue(ZosJobsProvider.persistenceSchema, setting);
             }
         }
     }